further cleanup types
