+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@lightningtv/core",
-  "version": "2.2.0-beta.2",
+  "version": "2.2.0-beta.3",
   "description": "Lightning TV Core for Universal Renderers",
   "type": "module",
   "main": "./dist/src/index.js",
